+++ .travis.yml
@@ -36,7 +36,7 @@ env:
   matrix:
     - "GEM=railties"
     - "GEM=ap,ac"
-    - "GEM=am,amo,as,av,aj,ast"
+    - "GEM=am,amo,as,av,aj"
     - "GEM=as PRESERVE_TIMEZONES=1"
     - "GEM=ar:mysql2"
     - "GEM=ar:sqlite3"
