+++ Gemfile
@@ -16,6 +16,8 @@ gem "rake", ">= 11.1"
 # be loaded after loading the test library.
 gem "mocha", "~> 0.14", require: false
 
+gem "capybara", "~> 2.7.0"
+
 gem "rack-cache", "~> 1.2"
 gem "jquery-rails"
 gem "coffee-rails"
