+++ actionpack/lib/action_dispatch.rb
@@ -97,6 +97,8 @@ module ActionDispatch
     autoload :TestResponse
     autoload :AssertionResponse
   end
+
+  autoload :SystemTestCase, "action_dispatch/system_test_case"
 end
 
 autoload :Mime, "action_dispatch/http/mime_type"
