+++ actionpack/test/dispatch/system_testing/browser_test.rb
@@ -1,9 +1,9 @@
-require "active_support/testing/autorun"
-require "action_system_test"
+require "abstract_unit"
+require "action_dispatch/system_testing/browser"
 
 class BrowserTest < ActiveSupport::TestCase
   test "initializing the browser" do
-    browser = ActionSystemTest::Browser.new(:chrome, [ 1400, 1400 ])
+    browser = ActionDispatch::SystemTesting::Browser.new(:chrome, [ 1400, 1400 ])
     assert_equal :chrome, browser.instance_variable_get(:@name)
     assert_equal [ 1400, 1400 ], browser.instance_variable_get(:@screen_size)
   end
