+++ actionpack/test/dispatch/system_testing/driver_test.rb
@@ -1,9 +1,9 @@
-require "active_support/testing/autorun"
-require "action_system_test"
+require "abstract_unit"
+require "action_dispatch/system_testing/driver"
 
 class DriverTest < ActiveSupport::TestCase
   test "initializing the driver" do
-    driver = ActionSystemTest::Driver.new(:selenium)
+    driver = ActionDispatch::SystemTesting::Driver.new(:selenium)
     assert_equal :selenium, driver.instance_variable_get(:@name)
   end
 end
