+++ actionpack/test/dispatch/system_testing/screenshot_helper_test.rb
@@ -1,9 +1,9 @@
-require "active_support/testing/autorun"
-require "action_system_test"
+require "abstract_unit"
+require "action_dispatch/system_testing/test_helpers/screenshot_helper"
 
 class ScreenshotHelperTest < ActiveSupport::TestCase
   test "image path is saved in tmp directory" do
-    new_test = ActionSystemTest::Base.new("x")
+    new_test = ActionDispatch::SystemTestCase.new("x")
 
     assert_equal "tmp/screenshots/failures_x.png", new_test.send(:image_path)
   end
