--- actionsystemtest/CHANGELOG.md
@@ -1,3 +1,0 @@
-*   Added to Rails!
-
-    *Eileen M. Uchitelle*
