--- actionsystemtest/lib/action_system_test/driver.rb
@@ -1,16 +1,0 @@
-module ActionSystemTest
-  class Driver # :nodoc:
-    def initialize(name)
-      @name = name
-    end
-
-    def run
-      register
-    end
-
-    private
-      def register
-        Capybara.default_driver = @name
-      end
-  end
-end
