--- actionsystemtest/lib/action_system_test/gem_version.rb
@@ -1,15 +1,0 @@
-module ActionSystemTest
-  # Returns the version of the currently loaded Action System Test as a <tt>Gem::Version</tt>.
-  def self.gem_version
-    Gem::Version.new VERSION::STRING
-  end
-
-  module VERSION
-    MAJOR = 5
-    MINOR = 1
-    TINY  = 0
-    PRE   = "alpha"
-
-    STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
-  end
-end
