+++ railties/lib/rails/generators/test_unit/system/templates/system_test.rb
@@ -1,7 +1,9 @@
 require "system_test_helper"
 
-class <%= class_name.pluralize %>Test < ActionSystemTestCase
-  # test 'the truth' do
-  #   assert true
+class <%= class_name.pluralize %>Test < ApplicationSystemTestCase
+  # test "visiting the index" do
+  #   visit <%= plural_table_name %>_url
+  #
+  #   assert_selector "h1", text: "<%= class_name %>"
   # end
 end
